Added tests for move, fixed some issue.
