+++ src/test/java/usecase_playeractions_tests/MoverTest.java
@@ -1,4 +1,4 @@
-package usecase_event_tests;
+package usecase_playeractions_tests;
 
 import entity.*;
 import org.junit.jupiter.api.Assertions;
@@ -6,39 +6,45 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import usecase_event.ArtifactEvent;
+import usecase_event.WallEvent;
+import usecase_playeractions.Map;
+import usecase_playeractions.Mover;
 
+import java.util.Observable;
 
-public class ArtifactEventTest {
+public class MoverTest {
+
+    Map map;
+    Player player;
     int maxHP = 100;
     int atkPt = 10;
     Collectible essence = new Collectible("Essence", 100);
-    Collectible artifact = new Collectible("Artifact", 1);
+    Collectible artifact = new Collectible("Artifact", 0);
     CollectibleInventory inventory = new CollectibleInventory("Collectible Inventory", essence, artifact);
     Armor armor = new Armor("Chain Mail", 5);
     Weapon excalibur = new Weapon("Legendary Sword Excalibur", 1000);
     BasicEquipmentSlots equipmentSlots = new BasicEquipmentSlots(excalibur, armor);
     int[] location = new int[]{0, 0};
-    Player player;
-    ArtifactEvent event;
-
 
 
     @BeforeEach
     @DisplayName("Setup before Each Test")
     void setUp(){
-        event = new ArtifactEvent();
+        map = new Map();
         player = new Player(maxHP, atkPt, inventory, equipmentSlots, location);
     }
 
     @Test
-    @DisplayName("Test Trigger")
-    void testTrigger(){
-        event.trigger(player);
-        Assertions.assertEquals(2, player.getCollectible("Artifact").getNum());
+    @DisplayName("Test Move")
+    void testMove(){
+        map.setBoard(new ArtifactEvent(),0,1);
+        Mover mover = new Mover(player,map,"W",0,1);
+        mover.update(new Observable(),"W");
+        Assertions.assertEquals(1, player.getPlayerLocation()[1]);
+        Assertions.assertEquals(1, player.getCollectible("Artifact").getNum());
+        map.setBoard(new WallEvent(),0,2);
+        mover.update(new Observable(),"W");
+        Assertions.assertEquals(1, player.getPlayerLocation()[1]);
     }
-    @Test
-    @DisplayName("Test Enter")
-    void testEnter(){
-        Assertions.assertTrue(event.enter(player));
-    }
+
 }
